+++ lib/ReactViews/StandardUserInterface/MapColumn.jsx
@@ -20,9 +20,23 @@ import Loader from "../Loader";
 import Styles from "./map-column.scss";
 import { observer } from "mobx-react";
 import SlideUpFadeIn from "../Transitions/SlideUpFadeIn/SlideUpFadeIn";
+import Button from "../../Styled/Button";
 
 const isIE = FeatureDetection.isInternetExplorer();
 const chromeVersion = FeatureDetection.chromeVersion();
+
+const flags = {
+  moveForward: false,
+  moveBackward: false,
+  moveUp: false,
+  moveDown: false,
+  moveLeft: false,
+  moveRight: false,
+  rotateUp: false,
+  rotateDown: false,
+  rotateLeft: false,
+  rotateRight: false
+};
 
 /**
  * Right-hand column that contains the map, controls that sit over the map and sometimes the bottom dock containing
@@ -89,6 +103,21 @@ const MapColumn = observer(
       }
     },
 
+    onKeyDown(event) {
+      const flagName = getFlagForKeyCode(event.key);
+      if (typeof flagName !== "undefined") {
+        flags[flagName] = true;
+        keyboardTickFunc(this.props);
+      }
+    },
+
+    onKeyUp(event) {
+      const flagName = getFlagForKeyCode(event.key);
+      if (typeof flagName !== "undefined") {
+        flags[flagName] = false;
+      }
+    },
+
     render() {
       const { customElements } = this.props;
       // const { t } = this.props;
@@ -98,11 +127,15 @@ const MapColumn = observer(
       const mapCellClass = classNames(Styles.mapCell, {
         [Styles.mapCellChrome]: isAboveChrome75
       });
+      const keyboardControlDescription =
+        "Navigazione della mappa da tastiera\n  w = zoom in\n  s = zoom out\n  q = muovi in su\n  e = muovi in giù\n  d = muovi a sinistra\n  a = muovi a destra\n  r = ruota in su\n  f = ruota in giù\n  z = ruota a sinistra\n  x = ruota a destra";
       return (
         <div
           className={classNames(Styles.mapInner, {
             [Styles.mapInnerChrome]: isAboveChrome75
           })}
+          onKeyDown={this.onKeyDown}
+          onKeyUp={this.onKeyUp}
         >
           <div className={Styles.mapRow}>
             <div
@@ -196,6 +229,30 @@ const MapColumn = observer(
                   <div key={i}>{feedbackItem}</div>
                 </For>
               </If>
+              <If
+                condition={
+                  !this.props.viewState.useSmallScreenInterface &&
+                  !this.props.viewState.hideMapUi
+                }
+              >
+                <Button
+                  type="button"
+                  title={keyboardControlDescription}
+                  style={{
+                    right: "20px",
+                    bottom: this.props.terria.configParameters.feedbackUrl
+                      ? "110px"
+                      : "75px",
+                    position: "absolute"
+                  }}
+                  rounded
+                  shortMinHeight
+                  styledMinWidth={"80px"}
+                  tabIndex={1}
+                >
+                  Naviga da tastiera
+                </Button>
+              </If>
             </div>
             <If condition={this.props.terria.configParameters.printDisclaimer}>
               <div className={classNames(Styles.mapCell, "print")}>
@@ -224,5 +281,74 @@ const MapColumn = observer(
     }
   })
 );
+
+function getFlagForKeyCode(keyCode) {
+  switch (keyCode) {
+    case "w":
+      return "moveForward";
+    case "s":
+      return "moveBackward";
+    case "q":
+      return "moveUp";
+    case "e":
+      return "moveDown";
+    case "d":
+      return "moveRight";
+    case "a":
+      return "moveLeft";
+    case "r":
+      return "rotateUp";
+    case "f":
+      return "rotateDown";
+    case "z":
+      return "rotateLeft";
+    case "x":
+      return "rotateRight";
+    default:
+      return undefined;
+  }
+}
+
+function keyboardTickFunc(props) {
+  const scene = props.terria.cesium.scene;
+  const ellipsoid = scene.globe.ellipsoid;
+  const camera = scene.camera;
+  const cameraHeight = ellipsoid.cartesianToCartographic(camera.position)
+    .height;
+  const moveRate = cameraHeight / 100.0;
+
+  if (flags.moveForward) {
+    camera.moveForward(moveRate);
+  }
+  if (flags.moveBackward) {
+    camera.moveBackward(moveRate);
+  }
+  if (flags.moveUp) {
+    camera.moveUp(moveRate);
+  }
+  if (flags.moveDown) {
+    camera.moveDown(moveRate);
+  }
+  if (flags.moveLeft) {
+    camera.moveLeft(moveRate);
+  }
+  if (flags.moveRight) {
+    camera.moveRight(moveRate);
+  }
+  if (flags.rotateUp) {
+    camera.lookUp();
+  }
+  if (flags.rotateDown) {
+    camera.lookDown();
+  }
+  if (flags.rotateLeft) {
+    camera.lookLeft();
+  }
+  if (flags.rotateRight) {
+    camera.lookRight();
+  }
+
+  props.terria.currentViewer.notifyRepaintRequired();
+}
 
 export default withTranslation()(MapColumn);